дз4

+++ repository/index.js
@@ -1,17 +1,38 @@
 const Contact = require('../model/contact')
 
-const listContacts = async () => {
-  const results = await Contact.find({})
-  return results
+const listContacts = async (userId, query) => {
+  // const results = await Contact.find({ owner: userId }).populate({
+  //   path: 'owner',
+  //   select: 'email subscription createdAt updatedAt'
+  // })
+  const {favorite = null, limit = 3, page = 1 } = query
+  const searchOptions = { owner: userId }
+  if (favorite !== null) {
+    searchOptions.favorite = favorite
+  }
+  const results = await Contact.paginate(searchOptions, { 
+    limit, 
+    page, 
+    populate:{
+    path: 'owner',
+    select: 'email subscription createdAt updatedAt'
+  } 
+})
+  const {docs: contacts} = results
+  delete results.docs
+  return {...results, contacts}
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findById(contactId)
+const getContactById = async (contactId, userId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription createdAt updatedAt'
+  })
   return result
 }
 
 const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndRemove({ _id: contactId })
+  const result = await Contact.findOneAndRemove({ _id: contactId, owner: userId })
   return result
 }
 const addContact = async (body) => {
@@ -19,9 +40,9 @@ const addContact = async (body) => {
   return result
 }
 
-const updateContact = async (contactId, body) => {
-  const result = await Contact.findByIdAndUpdate(
-    contactId,
+const updateContact = async (contactId, body, userId) => {
+  const result = await Contact.findOneAndUpdate(
+    { _id: contactId, owner: userId },
     { ...body },
     { new: true },
   )
